refactor: 파일 및 폴더 구조변경 #149
