+++ src/app/concert/form/[id]/_shared/components/form-info/form-info.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import ConcertDetail from '@/shared/components/concert-detail/concert-detail';
+import ConcertDetail from '@/shared/components/features/concert/concert-detail/concert-detail';
 import Badge from '@/shared/components/ui/badge/badge';
 import {
   TICKET_SITE_URL_MAP,
