+++ src/shared/components/features/concert/concert-detail/concert-detail.tsx
